sync RMF to 5034df6feff9073f7d5746e63e975bebc57173b6
git-svn-id: https://svn.salilab.org/imp/trunk@15839 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/rmf/dependency/RMF/src/backend/hdf5/HDF5SharedData.h
@@ -836,7 +836,7 @@ namespace RMF {
        }
        void add_child_frame(int node, int child_node){}
        Ints get_children_frame(int node) const {
-         if (node != get_number_of_frames()-1) {
+         if (node != static_cast<int>(get_number_of_frames())-1) {
            return Ints(1, node+1);
          } else {
            return Ints();
